flask-import/requirements.txt
appnope==0.1.0
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
biopython==1.73
bleach==3.1.0
certifi==2019.6.16
Click==7.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
Flask==1.1.1
Flask-Cors==3.0.8
imbalanced-learn==0.4.3
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
neo4j==4.3.9
notebook==5.7.8
numpy==1.16.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-dotenv==0.20.0
pytz==2018.9
pyzmq==18.0.1
qtconsole==4.4.3
scikit-learn==0.21.2
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
terminado==0.8.2
testpath==0.4.2
torch==1.0.1.post2
torchvision==0.2.2.post3
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.3.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.4.2
wrapt==1.11.1
